Simplify texture indexing, add hybrid sampling
This hybrid sampling approach achieves a good balance between quality and
performance by using bilinear filtering for magnification, and
nearest + nearest mipmap for minification.

The duplicate row/col for boundless linear sampling was removed, as it was
pretty annoying to deal with when updating pixels and it made indexing
relatively expansive (mullo takes 10c, could use shift for 1c instead).

Clamping is only needed for bilinear sampling and it can be done cheaply.