Combine material textures into a single layered texture
The idea behind this is that it allows the compiler to CSE setup values
like wrapped and scaled UVs, mip-levels and pixel offsets after inlining
multiple sample calls with the UVs.

This also adds support for emissive maps, using a trick with the albedo
alpha channel to reduce the number of sample calls.